+++ app/models/boat_maintenance_part.ts
@@ -1,4 +1,5 @@
 import { BoatMaintenancePartSchema } from '#database/schema'
+import BoatEnginePart from '#models/boat_engine_part'
 import BoatMaintenanceEvent from '#models/boat_maintenance_event'
 import { belongsTo } from '@adonisjs/lucid/orm'
 import type { BelongsTo } from '@adonisjs/lucid/types/relations'
@@ -6,4 +7,7 @@ import type { BelongsTo } from '@adonisjs/lucid/types/relations'
 export default class BoatMaintenancePart extends BoatMaintenancePartSchema {
   @belongsTo(() => BoatMaintenanceEvent, { foreignKey: 'maintenanceEventId' })
   declare maintenanceEvent: BelongsTo<typeof BoatMaintenanceEvent>
+
+  @belongsTo(() => BoatEnginePart, { foreignKey: 'enginePartId' })
+  declare enginePart: BelongsTo<typeof BoatEnginePart>
 }